+++ api/src/certification/results/domain/usecases/get-session-results-by-result-recipient-email.js
